[pilot] Moving mock discovery to cmd package
Having it in the same directory with pilot/test/mock/config.go was
causing it to bring in the pilot/testing package, which causes the
list of default flags to explode (testing.*).
