+++ pilot/cmd/pilot-discovery/mock/service.go
